# Conflicts:
#	abstrakcyjne/Hero.h

+++ abstrakcyjne/AddHpEvent.h
@@ -14,7 +14,7 @@ public:
 	virtual void processEvent() {
 		Hero * hero = gamemodel->getHero();
 		int maxhp = hero->getMAX_HP();
-		hero->addHP(maxhp * 0.05);
+		hero->addHP((int)(maxhp * 0.05));
 		if (hero->getCURRENT_HP() > maxhp)hero->setCURRENT_HP(maxhp);
 		gamemodel->getEventQ()->addEvent(new AddHpEvent(gamemodel, graphiceventq, processingTime + 5000));
